fix:total enc

+++ HTMDv2.2c-f303/App/Inc/dc_motor_controller.hpp
@@ -67,5 +67,5 @@ public:
      * @brief エンコーダーのサンプリング
      *
      */
-    int16_t sample_encoder();
+    float sample_encoder();
 };
